스프링과 JPA 기반 웹 애플리케이션 개발 강좌 (모임 참가 신청 수락, 취소 및 출석 체크)

+++ src/main/java/com/ghsong/studymeeting/event/EventService.java
@@ -51,8 +51,26 @@ public class EventService {
 
     public void cancelEnrollment(Event event, Account account) {
         Enrollment enrollment = enrollmentRepository.findByEventAndAccount(event, account);
-        event.removeEnrollment(enrollment);
-        enrollmentRepository.delete(enrollment);
-        event.acceptNextWaitingEnrollment();
+        if (!enrollment.isAttended()) {
+            event.removeEnrollment(enrollment);
+            enrollmentRepository.delete(enrollment);
+            event.acceptNextWaitingEnrollment();
+        }
+    }
+
+    public void acceptEnrollment(Event event, Enrollment enrollment) {
+        event.accept(enrollment);
+    }
+
+    public void reject(Event event, Enrollment enrollment) {
+        event.reject(enrollment);
+    }
+
+    public void checkin(Event event, Enrollment enrollment) {
+        event.checkin(enrollment);
+    }
+
+    public void cancelCheckin(Event event, Enrollment enrollment) {
+        event.cancelCheckin(enrollment);
     }
 }
